+++ scripts/create-container.js
@@ -4,13 +4,15 @@ const humps = require('humps');
 const logger = require('rear-logger')('create-container');
 const checkDir = require('./lib/check-dir');
 const isPath = require('./lib/is-path');
+const parseArgs = require('./lib/parse-args');
 const createTemplates = require('./lib/create-templates');
+const styleExt = require('./lib/style-extension');
 
 const CONTAINER_TEMPLATE = `
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { getProps, getActions } from './Providers';
-import './$1.css';
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import { getProps, getActions } from "./Providers";
+import "./$1.${styleExt}";
 
 class $1 extends Component {
 	render() {
@@ -24,8 +26,24 @@ class $1 extends Component {
 export default connect(getProps, getActions)($1);
 `;
 
+const CONTAINER_TEMPLATE_LIGHT = `
+import React, { Component } from "react";
+import "./$1.${styleExt}";
+
+class $1 extends Component {
+  render() {
+    return(
+      <div className="$1">
+      </div>
+    );
+  }
+}
+
+export default $1;
+`;
+
 const INDEX_TEMPLATE = `
-export { default } from './$1';
+export { default } from "./$1";
 `
 const CSS_TEMPLATE = `
 .$1 {
@@ -34,7 +52,7 @@ const CSS_TEMPLATE = `
 `;
 
 const PROVIDER_TEMPLATE = `
-// import { myAction } from '../../actions/MyActions';
+// import { myAction } from "../../actions/MyActions";
 
 export function getActions(dispatch) {
   return {
@@ -57,7 +75,11 @@ if (require.main === module) {
 
 /////////////////////////////////
 
-function Main(args) {
+function Main(argv) {
+  const program = parseArgs(argv);
+  const { isLight } = program.options;
+  const { args } = program;
+
   const containersPath = isPath(args[0])
     ? args.shift(0, 1)
     : path.resolve(__dirname, '..', 'src', 'containers');
@@ -68,11 +90,13 @@ function Main(args) {
 
   if (!checkDir(containerDir)) return;
 
-  createTemplates([{
+  const TEMPLATE = isLight ? CONTAINER_TEMPLATE_LIGHT : CONTAINER_TEMPLATE;
+
+  const files = [{
     file: path.join(containerDir, `${containerName}.js`),
-    data: CONTAINER_TEMPLATE.replace(PATTERN, containerName)
+    data: TEMPLATE.replace(PATTERN, containerName)
   }, {
-    file: path.join(containerDir, `${containerName}.css`),
+    file: path.join(containerDir, `${containerName}.${styleExt}`),
     data: CSS_TEMPLATE.replace(PATTERN, containerName)
   }, {
     file: path.join(containerDir, `Providers.js`),
@@ -80,7 +104,12 @@ function Main(args) {
   }, {
     file: path.join(containerDir, 'index.js'),
     data: INDEX_TEMPLATE.replace(PATTERN, containerName)
-  }]);
+  }];
+
+  // removes provider template
+  if (isLight) files.splice(2, 1);
+
+  createTemplates(files);
 
   logger.log();
 }